inquire-app/inquire/api/permissions.py
from rest_framework import permissions
from rest_framework.permissions import IsAuthenticatedOrReadOnly, IsAuthenticated


class IsOwnerOrReadOnly(IsAuthenticatedOrReadOnly):
    def has_object_permission(self, request, view, obj):
        if request.method in permissions.SAFE_METHODS:
            return True
        return obj.profile == request.user


class IsOwner(IsAuthenticated):
    def has_object_permission(self, request, view, obj):
        return obj.profile == request.user
