inquire-app/requirements.txt
Django==1.11.4
djangorestframework==3.6.4
pytz==2017.2
psycopg2==2.7.3.1
requests==2.18.4